Started working on Chest opening

+++ src/main/java/dev/dbdh/Discord/Listeners/Fun/Chest.java
@@ -10,23 +10,26 @@ import java.time.Instant;
 
 public class Chest extends ListenerAdapter {
 
-    public void onGuildMessageReceived(GuildMessageReceivedEvent event){
+    public void onGuildMessageReceived(GuildMessageReceivedEvent event) {
         String[] args = event.getMessage().getContentRaw().split("\\s+");
         Color color = new Color();
         Data data = new Data();
         EmbedBuilder eb = new EmbedBuilder();
-        if(args[0].equalsIgnoreCase(data.getPrefix() + "chest")){
-            if(args.length < 2){
-                eb.setDescription("You need to specify what type of chest you would like to open. `Basic | Shiny | Legendary | Mythic`");
-                eb.setColor(color.getRandomColor());
-                eb.setTimestamp(Instant.now());
-                eb.setFooter("Insufficient Arguments", data.getSelfAvatar(event));
+        if (args[0].equalsIgnoreCase(data.getPrefix() + "chest")) {
+            if (event.getMessage().getChannel().equals(event.getGuild().getTextChannelById("632350945891581992"))) {
+                if (args.length < 2) {
+                    eb.setDescription("You need to specify what type of chest you would like to open. `Basic | Shiny | Legendary | Mythic`");
+                    eb.setColor(color.getRandomColor());
+                    eb.setTimestamp(Instant.now());
+                    eb.setFooter("Insufficient Arguments", data.getSelfAvatar(event));
 
-                event.getChannel().sendMessage(eb.build()).queue((message) -> {
-                    message.delete().queue();
-                });
+                    event.getChannel().sendMessage(eb.build()).queue((message) -> {
+                        message.delete().queue();
+                    });
+                }
+            } else if (args.length == 2) {
+
             }
         }
     }
-
 }
